documentation & cleanup
